requirements.txt
# Core ML
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.37.0
open-clip-torch>=2.24.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Vector search
faiss-cpu>=1.7.4

# OCR
easyocr>=1.7.1

# API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
aiofiles>=23.2.1

# Frontend
streamlit>=1.31.0

# Data
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.4.0
tqdm>=4.66.0

# Config
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Knowledge enrichment
wikipedia>=1.4.0
SPARQLWrapper>=2.0.0
requests>=2.31.0

# Reddit
praw>=7.7.1

# Database
# sqlite3 is part of stdlib

# Evaluation
scipy>=1.12.0

# Utilities
httpx>=0.26.0
